+++ Extensions/WriteableBitmapExtensions.cs
@@ -2,7 +2,7 @@
 using System.Windows;
 using System.Windows.Media.Imaging;
 
-namespace Chaldea.Components.Extensions
+namespace TamamoToolkit.Extensions
 {
     /// <summary>
     /// <see cref="WriteableBitmap"/> 扩展类
@@ -23,14 +23,14 @@ namespace Chaldea.Components.Extensions
         /// 使用像素数组更新位图中的像素
         /// </summary>
         /// <param name="bitmap"></param>
-        /// <param name="source">用来更新位图的像素数数组</param>
+        /// <param name="source">用来更新位图的像素数组</param>
         /// <exception cref="NotImplementedException">像素数组为不支持的格式</exception>
         public static void WritePixels(this WriteableBitmap bitmap, Array source)
         {
             try
             {
                 byte[] bytes = new byte[bitmap.PixelWidth * bitmap.PixelHeight * bitmap.Format.BitsPerPixel / 8];
-                Buffer.BlockCopy(source, 0, bytes, 0, bytes.Length);
+                Buffer.BlockCopy(source, 0, bytes, 0, Math.Min(bytes.Length, Buffer.ByteLength(source)));
                 bitmap.Lock();
                 Marshal.Copy(bytes, 0, bitmap.BackBuffer, bytes.Length);
                 bitmap.AddDirtyRect(new Int32Rect(0, 0, bitmap.PixelWidth, bitmap.PixelHeight));
@@ -38,7 +38,7 @@ namespace Chaldea.Components.Extensions
             }
             catch (Exception ex)
             {
-                throw new NotImplementedException("像素数据为不支持的格式", ex);
+                throw new NotImplementedException("像素数组为不支持的格式", ex);
             }
         }
     }