+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-﻿## v2.0.0 更新内容
+﻿## v2.0.1 更新内容
+
+- 项目正式更名为 **TamamoToolkit.Windows**，所有命名空间均已调整
+	- Chaldea.Components -> TamamoToolkit
+	- Chaldea.Components.DllImport -> TamamoToolkit.DllImport
+	- Chaldea.Components.Extensions -> TamamoToolkit.Extensions
+	- Chaldea.Components.Utils -> TamamoToolkit.Utils
+
+## v2.0.0 更新内容
 
 - 首次发布，为确保一致性，故版本号从 v2.0.0 开始